makerworks/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
alembic
pydantic>=2.0
pydantic-settings
psycopg[binary]
redis
rq
httpx
strawberry-graphql
prometheus-client
